增加数据库加密

+++ CRTBTunnelMonitor/src/com/crtb/tunnelmonitor/utils/CrtbDbFileUtils.java
@@ -1,13 +1,13 @@
 package com.crtb.tunnelmonitor.utils;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.zw.android.framework.impl.ExecuteAsyncTaskImpl;
 
+import ICT.utils.DbAESEncrypt;
+import ICT.utils.IDbEncrypt;
 import android.content.Context;
 import android.os.Environment;
 
@@ -103,29 +103,21 @@ public final class CrtbDbFileUtils {
 				
 				String outPath = getExportPath(AppConfig.DB_EXPORT_DIR) + "/" + filename ;
 				
+				IDbEncrypt encrypt = new DbAESEncrypt() ;
+				
 				try {
-					FileInputStream in = new FileInputStream(path);
-					FileOutputStream out = new FileOutputStream(outPath);
 					
-					byte[] buffer = new byte[4 * 1024];
-					int read = -1 ;
-					
-					while((read = in.read(buffer)) > 0){
-						
-						AppLogger.d(TAG, "zhouwei : 导出数据库..."); 
-						
-						out.write(buffer, 0, read);
-						out.flush() ;
-					}
-					
-					in.close() ;
-					out.close() ;
+					// 加密
+					boolean success = encrypt.encrypt(path, outPath) ;
 					
 					Thread.sleep(2000);
 					
-					AppLogger.d(TAG, "zhouwei : 数据库导出完成: " + outPath); 
-					
-					sendMessage(MSG_EXPORT_DB_SUCCESS) ;
+					if(success){
+						AppLogger.d(TAG, "zhouwei : 数据库导出完成: " + outPath); 
+						sendMessage(MSG_EXPORT_DB_SUCCESS) ;
+					} else {
+						sendMessage(MSG_EXPORT_DB_FAILED) ;
+					}
 				} catch (Exception e1) {
 					e1.printStackTrace();
 					sendMessage(MSG_EXPORT_DB_FAILED) ;
@@ -150,43 +142,37 @@ public final class CrtbDbFileUtils {
 				String filename = path.substring(path.lastIndexOf("/") + 1);
 				String outPath 	= "data/data/" + context.getPackageName() + "/databases/" + filename ;
 				
+				IDbEncrypt encrypt = new DbAESEncrypt() ;
+				
 				try {
-					FileInputStream in 		= new FileInputStream(path);
-					FileOutputStream out 	= new FileOutputStream(outPath);
 					
-					byte[] buffer = new byte[4 * 1024];
-					int read = -1 ;
+					// 解密
+					boolean success = encrypt.decrypt(path, outPath) ;
 					
-					while((read = in.read(buffer)) > 0){
+					if(success){
 						
-						AppLogger.d(TAG, "zhouwei : 导入数据库..."); 
+						// 写入数据库
+						int code = ProjectIndexDao.defaultWorkPlanDao().inportDb(filename) ;
 						
-						out.write(buffer, 0, read);
-						out.flush() ;
-					}
-					
-					in.close() ;
-					out.close() ;
-					
-					// 写入数据库
-					int code = ProjectIndexDao.defaultWorkPlanDao().inportDb(filename) ;
-					
-					Thread.sleep(2000);
-					
-					if(code == 100){
+						Thread.sleep(2000);
 						
-						// 删除文件
-						File f = new File(outPath);
-						f.delete() ;
-						
-						AppLogger.d(TAG, "zhouwei : 数据库导入失败"); 
-						
-						sendMessage(MSG_INPORT_DB_FAILED,"非注册用户,不能保存2个以上的工作面") ;
+						if(code == 100){
+							
+							// 删除文件
+							File f = new File(outPath);
+							f.delete() ;
+							
+							AppLogger.d(TAG, "zhouwei : 数据库导入失败"); 
+							
+							sendMessage(MSG_INPORT_DB_FAILED,"非注册用户,不能保存2个以上的工作面") ;
+						} else {
+							
+							AppLogger.d(TAG, "zhouwei : 数据库导入完成: " + outPath); 
+							
+							sendMessage(MSG_INPORT_DB_SUCCESS) ;
+						}
 					} else {
-						
-						AppLogger.d(TAG, "zhouwei : 数据库导入完成: " + outPath); 
-						
-						sendMessage(MSG_INPORT_DB_SUCCESS) ;
+						sendMessage(MSG_INPORT_DB_FAILED) ;
 					}
 				} catch (Exception e1) {
 					e1.printStackTrace();
